Compare IDG
IDG seems to be incorrect on Takiddine Youssouf vs Djaffar Salim in
Youth/Sports etc. These two switched roles in the April reshuffle.

It's possible that Youssouf returned to this role since, but I haven't
been able to find anything about that, and later sources also confirm
Salim in the role.

[1] e.g. https://country.eiu.com/article.aspx?articleid=1812419164&Country=Comoros&topic=Summary&subtopic=P_8